shap_explain/explain_requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.41.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
jax==0.4.13
joblib==1.3.1
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.6
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
Pillow==10.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
shap==0.42.1
six==1.16.0
slicer==0.0.7
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.16.2
